+++ IoTDBJDBC/src/main/java/com/oracle/demo/timg/iot/iotdbjdbc/dataread/aq/IoTAQCore.java
@@ -1,4 +1,4 @@
-package com.oracle.demo.timg.iot.iotdbjdbc.dataread;
+package com.oracle.demo.timg.iot.iotdbjdbc.dataread.aq;
 
 import java.sql.CallableStatement;
 import java.sql.SQLException;
