+++ IoTDBJDBC/src/main/java/com/oracle/demo/timg/iot/iotdbjdbc/dataread/aq/IoTAQNormalizedDataBatchReader.java
@@ -1,10 +1,11 @@
-package com.oracle.demo.timg.iot.iotdbjdbc.dataread;
+package com.oracle.demo.timg.iot.iotdbjdbc.dataread.aq;
 
 import java.sql.SQLException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 import com.oracle.demo.timg.iot.iotdbjdbc.aqdata.NormalizedData;
+import com.oracle.demo.timg.iot.iotdbjdbc.dataread.IoTDBClient;
 import com.oracle.demo.timg.iot.iotdbjdbc.messagehandler.NormalizedDataMessageHandlerService;
 import com.oracle.demo.timg.iot.iotdbjdbc.oci.DBConnectionSupplier;
