+++ IoTDBJDBC/src/main/java/com/oracle/demo/timg/iot/iotdbjdbc/messagehandler/testtools/NormalizedDataDuplicatorMessageHandler.java
@@ -1,6 +1,7 @@
-package com.oracle.demo.timg.iot.iotdbjdbc.messagehandler;
+package com.oracle.demo.timg.iot.iotdbjdbc.messagehandler.testtools;
 
 import com.oracle.demo.timg.iot.iotdbjdbc.aqdata.NormalizedData;
+import com.oracle.demo.timg.iot.iotdbjdbc.messagehandler.NormalizedDataMessageHandler;
 
 import io.micronaut.context.annotation.Property;
 import io.micronaut.context.annotation.Requires;
