+++ IoTDBJDBC/src/main/java/com/oracle/demo/timg/iot/iotdbjdbc/runner/JDBCRunner.java
@@ -59,7 +59,7 @@ public class JDBCRunner {
 		this.ioTDBClients = ioTDBClients.stream().sorted().toList();
 		String iotDBClientDetails = this.ioTDBClients.stream()
 				.map(client -> client.getName() + "(" + client.getConfig() + ")").collect(Collectors.joining(", "));
-		log.info("There are " + ioTDBClients.size() + " IoTDBClients configured - " + iotDBClientDetails);
+		log.info("There are " + this.ioTDBClients.size() + " IoTDBClients configured - " + iotDBClientDetails);
 		// place to put the clients we have started
 		configuredClients = new ArrayList<>(ioTDBClients.size());
 	}
